add Gyo.noConflict() to prevent conflict with external library because of alias keyword.
