+++ index.html
@@ -5,8 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>gyobutton example</title>
-    <link rel="stylesheet" type="text/css" href="./src/css/index.css">
-    <script type="text/javascript" src="./dist/js/bundle.js"></script>
+    <link rel="stylesheet" type="text/css" href="./index.css">
+    <!-- <script type="text/javascript" src="./dist/js/gyobutton.js"></script> -->
+    <script type="text/javascript" src="./dist/js/gyobutton.js"></script>
 </head>
 <body>
     <button id="i-1" class="button">BUTTON</button>
@@ -19,6 +20,7 @@
     <button class="toggle-style-custom">Toggle Style Custom</button>
     <button class="toggle-style-custom">Toggle Style Custom</button>
     <button class="toggle-style-custom">Toggle Style Custom</button>
+    <button id="toggle-conflict">Toggle Style Custom</button>
 </body>
 </html>
 <script type="text/javascript">
@@ -65,4 +67,11 @@
     let g4 = new Gyo.toggleButton(".toggle-style-custom");
     g4.button();
     g4.toggle(onToggleCallback, {color:'red'}, {color:'black'})
+
+    let G = Gyo.noConflict(true);
+
+    console.log(G);
+    let g5 = new G.toggleButton("#toggle-conflict");
+    g5.button();
+    g5.toggle(onToggleCallback, {color:'green'}, {color:'black'})
 </script>
